requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.36
alembic==1.13.3
pydantic==2.9.2
python-multipart==0.0.12
requests==2.32.3
Pillow==10.4.0
pytesseract==0.3.13
opencv-python==4.10.0.84
numpy==2.1.2
pytest==8.3.3
httpx==0.27.2
